Adapt for mainline changes in https://github.com/HelenOS/helenos/pull/15, and create a toolchain facade with a `-helenos` target tuple.
Allows a simpler, more natural cross-compilation and platform support.
For most packages using autoconf, changes necessary to support this are
limited to replacing the `config.sub` library file. GCC and binutils are
configured to build compiler for the "real" target platform (this is called
"Canadian cross", --build, --host and --target are all different). As an
exception, PCC is build for `-linux-gnu` target. python2 builds with a small
configure hack, but there's a lot of error messages that end up being ignored
for some reason. Not sure if that's a regression or not.